Write an http proxy, try to avoid manual parsoid config.
Still getting 404s. Needs more debugging.

This is at least part of the puzzle though; if we set a config for
parsoid here, then the visual editor plugin won't enable it
automatically. If we don't, it insists on using $_SERVER['HTTP_HOST'] as
the domain, which means it will fail to connect if we do nothing else.
So this patch includes a trivial HTTP proxy that always connects to the
local nginx.

An alternative approach might be to just set up parsoid manually.